python/pyproject.toml
[project]
name = "agent-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.1",
    "langgraph>=0.2.66",
    "livekit-agents>=0.12.9",
    "livekit-plugins-deepgram>=0.6.16",
    "livekit-plugins-openai>=0.10.16",
    "livekit-plugins-silero>=0.7.4",
    "python-dotenv>=1.0.1",
]
